+++ app/Http/Controllers/ProfileController.php
@@ -6,7 +6,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\ApplicantProfileRequest;
 use App\Http\Requests\ProfileUpdateRequest;
-
+use Illuminate\Support\Facades\Storage;
 use App\User;
 use App\Photo;
 use App\Package;
@@ -47,7 +47,9 @@ class ProfileController extends Controller
                 ]);
                 $input['data'][$key] = $photo->id;
                 if ($user->photo($type)) {
-                    unlink(public_path() . $user->photo($type)->path);
+                    if (Storage::exists($user->photo($type)->path)) {
+                        unlink(public_path() . $user->photo($type)->path);
+                    }
                     Photo::findOrFail($user->photo($type)->id)->delete();
                 }
             }
@@ -65,7 +67,9 @@ class ProfileController extends Controller
 
             if ($user->photo) {
 
-                unlink(public_path() . $user->photo->path);
+                if (Storage::exists($user->photo->path)) {
+                    unlink(public_path() . $user->photo->path);
+                }
                 Photo::findOrFail($user->photo->id)->delete();
             }
         }
@@ -111,7 +115,9 @@ class ProfileController extends Controller
                 ]);
                 $input['data'][$key] = $photo->id;
                 if ($user->photo($type)) {
-                    unlink(public_path() . $user->photo($type)->path);
+                    if (Storage::exists($user->photo($type)->path)) {
+                        unlink(public_path() . $user->photo($type)->path);
+                    }
                     Photo::findOrFail($user->photo($type)->id)->delete();
                 }
             }
@@ -130,7 +136,9 @@ class ProfileController extends Controller
 
             if ($user->photo) {
 
-                unlink(public_path() . $user->photo->path);
+                if (Storage::exists($user->photo->path)) {
+                    unlink(public_path() . $user->photo->path);
+                }
                 Photo::findOrFail($user->photo->id)->delete();
             }
         }
@@ -195,7 +203,9 @@ class ProfileController extends Controller
                 ]);
                 $input['data'][$key] = $photo->id;
                 if ($user->photo($type)) {
-                    unlink(public_path() . $user->photo($type)->path);
+                    if (Storage::exists($user->photo($type)->path)) {
+                        unlink(public_path() . $user->photo($type)->path);
+                    }
                     Photo::findOrFail($user->photo($type)->id)->delete();
                 }
             }
@@ -210,7 +220,9 @@ class ProfileController extends Controller
             ]);
             $input['photo_id'] = $photo->id;
             if ($user->photo) {
-                unlink(public_path() . $user->photo->path);
+                if (Storage::exists($user->photo->path)) {
+                    unlink(public_path() . $user->photo->path);
+                }
                 Photo::findOrFail($user->photo->id)->delete();
             }
         }
@@ -264,12 +276,16 @@ class ProfileController extends Controller
         }
         if ($user->invoiceProof) {
             $proof = Photo::findOrFail($user->invoiceProof->id);
-            unlink(public_path() . $proof->path);
+            if (Storage::exists($proof->path)) {
+                unlink(public_path() . $proof->path);
+            }
             $user->invoiceProof->delete();
         }
         if ($user->photo) {
             $photo = Photo::findOrFail($user->photo->id);
-            unlink(public_path() . $photo->path);
+            if (Storage::exists($photo->path)) {
+                unlink(public_path() . $photo->path);
+            }
             $user->photo->delete();
         }
