+++ app/Http/Controllers/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\Request;
 use App\Http\Requests\UsersCreateRequest;
 use App\Http\Requests\UsersUpdateRequest;
@@ -137,7 +138,9 @@ class UserController extends Controller
 
             if ($user->photo) {
 
-                unlink(public_path() . $user->photo->path);
+                if (Storage::exists($user->photo->path)) {
+                    unlink(public_path() . $user->photo->path);
+                }
                 Photo::findOrFail($user->photo->id)->delete();
             }
         }
@@ -160,7 +163,9 @@ class UserController extends Controller
     {
         $user = User::findOrFail($id);
         if ($user->photo) {
-            unlink(public_path() . $user->photo->path);
+            if (Storage::exists($user->photo->path)) {
+                unlink(public_path() . $user->photo->path);
+            }
             Photo::findOrFail($user->photo->id)->delete();
         }
         $user->delete();
